get rid of platform specific decode files
